changed unwarranted size_t to u16

+++ perfaware/sim86/shared/Sim8086.cpp
@@ -50,7 +50,7 @@ enum Flags
 static void SetFlags(const instruction& Instruction, const u16 LeftOperandValue, 
     const u16 RightOperandValue, const u16 Result, bool* FlagArray)
 {   
-    size_t HighOrderBit = (Instruction.Operands[1].Register.Count == 1) ? 0x80 : 0x8000;
+    u16 HighOrderBit = (Instruction.Operands[1].Register.Count == 1) ? 0x80 : 0x8000;
     bool Parity = true;
        
     switch (Instruction.Op)
